requirements.txt
numpy
opencv-python
transformers
diffusers==0.19.3
accelerate
torch-ema
imageio
imageio-ffmpeg
path
tqdm
Pillow
oflibpytorch
tensorboardX
pandas
rich
matplotlib
scikit-image
basicsr>=1.4.2
facexlib>=0.2.5
gfpgan>=1.3.5
detectron2
